add: page3 / communication
